move UltravoxDataproc back to model/ dir
