+++ ultravox/training/train.py
@@ -17,7 +17,7 @@ import wandb
 from torch.distributed.elastic.multiprocessing.errors import record
 from torch.utils import data
 
-from ultravox.data import data_processing
+from ultravox.model import data_processing
 from ultravox.data import datasets
 from ultravox.inference import infer
 from ultravox.model import ultravox_config
